feat: Add @rocicorp/zero/internal package
The idea behind this package is to provide a way to access the internal
APIs of Zero without having them get in the face of users who don't
need them. The main target for this export is for people implementing
solid (etc) bindings.
