+++ packages/zero-internal/src/mod.ts
@@ -1,0 +1,6 @@
+export type {Change} from '../../zql/src/zql/ivm/change.js';
+export type {Input, Output} from '../../zql/src/zql/ivm/operator.js';
+export {applyChange} from '../../zql/src/zql/ivm/view-apply-change.js';
+export type {Entry, Format, View} from '../../zql/src/zql/ivm/view.js';
+export type {Query, QueryType, Smash} from '../../zql/src/zql/query/query.js';
+export type {TableSchema} from '../../zql/src/zql/query/schema.js';
