+++ packages/zero/tool/build.js
@@ -57,6 +57,7 @@ async function getExternal(includePeerDeps) {
     'zero-protocol',
     'zero-react',
     'zero-solid',
+    'zero-internal',
     'zql',
     'zqlite',
   ]) {
@@ -96,6 +97,7 @@ async function buildZeroClient() {
     zero: basePath('src/zero.ts'),
     react: basePath('src/react.ts'),
     solid: basePath('src/solid.ts'),
+    internal: basePath('src/internal.ts'),
   };
   await esbuild.build({
     ...sharedOptions(false, false),
